move DeviceSize to its own directory
